+++ src/effective/core.clj
@@ -1,9 +1,9 @@
 (ns effective.core
-  (:require [effective.assertion :as assertion]
+  (:require [flatland.useful.fn :as ufn]
+            [effective.assertion :as assertion]
             [effective.assertion.composer :as composer]
             [effective.checkpoint :as checkpoint]
-            [effective.config :as config]
-            [effective.validation :as validation]))
+            [effective.config :as config]))
 
 (defmacro expect
   "Asserts modifications specified by `config` using the `clojure.test` API.
@@ -153,18 +153,22 @@
   ([effect config]
    `(expect ~effect :all ~config))
   ([effect composition config]
-   (if (validation/config-valid? config)
+   (if (config/valid? config)
      (let [observables-seq (config/observables config)
            before (interleave (map checkpoint/before (range)) observables-seq)
            after (interleave (map checkpoint/after (range)) observables-seq)
            composer (composer/make composition)
            join-intra (composer/intra composer)
            join-inter (composer/inter composer)
-           assertions (->> (map assertion/make config (range))
-                           (map join-intra)
+           assertions (->> config
+                           (config/groom)
+                           (map-indexed (comp join-intra
+                                              (ufn/ap assertion/make)
+                                              reverse
+                                              vector))
                            (join-inter))]
        `(let [~@before
               _# ~effect
               ~@after]
           ~@assertions))
-     `(throw (IllegalArgumentException. "Invalid configuration")))))
+     `(throw (IllegalArgumentException. ~(str (config/errors config)))))))